Folder created, archive directions corrected
